got rid of old main activity and all tab-related components
