+++ app/src/main/java/ca/chronofit/chrono/SplashActivity.kt
@@ -8,7 +8,6 @@ import android.view.View
 import androidx.databinding.DataBindingUtil
 import ca.chronofit.chrono.databinding.ActivitySplashBinding
 import ca.chronofit.chrono.main.MainActivity
-import ca.chronofit.chrono.main.NavigationActivity
 import ca.chronofit.chrono.util.BaseActivity
 
 private const val SPLASH_TIMEOUT = 3000L // Timeout delay
@@ -34,7 +33,7 @@ class SplashActivity : BaseActivity() {
             Looper.getMainLooper()
         ).postDelayed(
             {
-                startActivity(Intent(this, NavigationActivity::class.java))
+                startActivity(Intent(this, MainActivity::class.java))
                 finish()
             }, SPLASH_TIMEOUT
         )
