--- app/src/main/res/drawable/nav_selected.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="27dp" />
-    <stroke
-        android:width="1dp"
-        android:color="@color/colorAccent" />
-    <solid android:color="@color/colorAccent" />
-
-</shape>
