--- app/src/main/res/drawable/nav_tab_dark.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="27dp" />
-    <stroke
-        android:width="1dp"
-        android:color="@color/dark_grey" />
-    <solid android:color="@color/dark_grey" />
-
-</shape>
-
