+++ app/src/main/res/layout/fragment_settings.xml
@@ -7,6 +7,13 @@
         android:layout_height="match_parent"
         android:orientation="vertical">
 
+        <com.google.android.material.textview.MaterialTextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:gravity="center"
+            android:paddingTop="@dimen/xl_large_element_padding"
+            android:text="@string/todo"
+            android:textSize="@dimen/xl_large_element_padding" />
 
     </LinearLayout>
 </layout>
